+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Removed:
 - `@mappersmith/core`: Remove `middlewares` alias; use `middleware` instead.
 
 Changed:
+- `@mappersmith/core`: `Request.prototype.body` now respects `content-type` header and returns a JSON parsed result if the `content-type` is in the `+json` family. If you want the raw body without transformation, use `Request.prototype.rawBody` instead.
 - `@mappersmith/core`: No longer accepts `Promise` configuration, instead relying on native  Promise implementation to be available.
 - `@mappersmith/core`: Stop using `window.btoa` and use internal implementation instead.
 - `@mappersmith/*`: The project no longer produces a web artefact.